requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Pydantic
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Security & Crypto
cryptography==43.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # ← CAMBIAR: Fijar versión compatible
argon2-cffi==23.1.0

# Redis
redis==5.0.8

# Rate Limiting
slowapi==0.1.9

# Monitoring & Logging
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0

# HTTP Client
httpx==0.27.0

# Utilities
python-dotenv==1.0.1